[REMOVE] useless primitive cast

+++ src/main/java/com/github/unchama/seichiassist/task/AssaultTask.java
@@ -272,7 +272,7 @@ public class AssaultTask extends BukkitRunnable{
 		//減る経験値計算
 		//実際に破壊するブロック数 * 全てのブロックを破壊したときの消費経験値÷すべての破壊するブロック数 * 重力
 
-		double useMana = (double)breaksum * (double) (gravity + 1)
+		double useMana = (double)breaksum * (gravity + 1)
 				* ActiveSkill.getActiveSkillUseExp(playerdata.getActiveskilldata().assaulttype, playerdata.getActiveskilldata().assaultnum)
 				/(ifallbreaknum) ;
